fixed xp bar to render when activity checked
