+++ GamePlan.Web/src/Components/Navbar.jsx
@@ -5,9 +5,9 @@ import NavbarBackground from "../assets/images/Background_main.png";
 import Week from "./Week";
 import { WeekProvider, WeekContext } from "./WeekContext";
 
-
 const Navbar = (props) => {
   const [goalXp, setGoalXp] = useState(200);
+  // const [toggle, setToggle] = useState(false);
   const [user, setUser] = useState({});
   const {
     weekNumber,
@@ -15,6 +15,8 @@ const Navbar = (props) => {
     decreaseWeekNumber,
     month,
     setMonth,
+    toggleActivity,
+    nikos,
   } = useContext(WeekContext);
 
   const viewMonth = () => {
@@ -32,21 +34,27 @@ const Navbar = (props) => {
   };
 
   useEffect(() => {
+    user.xp = nikos.xp;
+  }, [toggleActivity]);
+
+  useEffect(() => {
     getUser();
     console.log("using effect");
   }, []);
 
   return (
     <nav className="navbar">
-      <div className="navbarLeft">
-        {/* <BurgerMenu /> */}
-      </div>
+      <div className="navbarLeft">{/* <BurgerMenu /> */}</div>
       <div className="navbarCenter">
         <h2 className="viewMonth">{month}</h2>
         <div className="navbarWeeks">
-          <button className="btn-weekNumber" onClick={decreaseWeekNumber}>Föreg.</button>
+          <button className="btn-weekNumber" onClick={decreaseWeekNumber}>
+            Föreg.
+          </button>
           <span className="weekNumber"> Vecka {weekNumber}</span>
-          <button className="btn-weekNumber" onClick={increaseWeekNumber}>Nästa</button>
+          <button className="btn-weekNumber" onClick={increaseWeekNumber}>
+            Nästa
+          </button>
         </div>
         <div className="goal" style={xpBar(user.xp / 2)}>
           <p>
